More fixture unscrambling

+++ tests/utils/__init__.py
@@ -1,7 +1,7 @@
 from math import log
 
 
-def approx(x1: int, x2: int, precision: int, abs_precision=None):
+def approx(x1: int, x2: int, precision: float, abs_precision=None):
     if precision >= 1:
         return True
     result = False
